NaN values again. Running smoke test for phase 1 expA

+++ scripts/experiments/exp1.py
@@ -31,7 +31,7 @@ MODEL_NAME = "google/gemma-2b"
 
 def _load_model() -> HookedTransformer:
     device = "cuda" if torch.cuda.is_available() else "cpu"
-    dtype = torch.float16 if device == "cuda" else torch.float32
+    dtype = torch.bfloat16 if device == "cuda" else torch.float32
     model = HookedTransformer.from_pretrained(
         MODEL_NAME,
         device=device,
